+++ benchmark_load_mov_metadata_test.go
@@ -9,7 +9,7 @@ import "github.com/amarburg/go-lazyfs-testfiles"
 
 var SparseStoreRoot = "cache/sparse/"
 
-type frameFunc func() int
+type frameFunc func() uint64
 
 func loadMovMetadataTask(b *testing.B, source lazyfs.FileSource, f frameFunc) {
 	b.ResetTimer()
@@ -34,7 +34,7 @@ func BenchmarkLoadMovMetadataFromLocalSource(b *testing.B) {
 		panic("Couldn't open Test file")
 	}
 
-	loadMovMetadataTask(b, fileSource, func() int { return 2 })
+	loadMovMetadataTask(b, fileSource, func() uint64 { return 2 })
 }
 
 func BenchmarkLoadMovMetadataFromLocalSourceSparseStore(b *testing.B) {
@@ -50,5 +50,5 @@ func BenchmarkLoadMovMetadataFromLocalSourceSparseStore(b *testing.B) {
 		panic("Couldn't open SparesFileFSStore")
 	}
 
-	loadMovMetadataTask(b, sparseStore, func() int { return 2 })
+	loadMovMetadataTask(b, sparseStore, func() uint64 { return 2 })
 }
